+++ testing/experimental/vector/push_back.cpp
@@ -56,7 +56,7 @@ void test_push_back()
     assert(result == copied_value);
     assert(v.size() == num_initial_elements + 1);
     assert(std::count(v.end()-1, v.end(), copied_value) == 1);
-    assert(std::count(v.begin(), v.end()-1, initial_value) == num_initial_elements);
+    assert(std::count(v.begin(), v.end()-1, initial_value) == static_cast<int>(num_initial_elements));
   }
 
   {
@@ -76,7 +76,7 @@ void test_push_back()
     assert(result == moved_value);
     assert(v.size() == num_initial_elements + 1);
     assert(std::count(v.end()-1, v.end(), moved_value) == 1);
-    assert(std::count(v.begin(), v.end()-1, initial_value) == num_initial_elements);
+    assert(std::count(v.begin(), v.end()-1, initial_value) == static_cast<int>(num_initial_elements));
   }
 }
